django_project/requirements.txt
boto3==1.10.35
botocore==1.13.35
certifi==2018.10.15
chardet==3.0.4
Django==2.1
django-crispy-forms==1.7.2
django-restframework==0.0.1
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
idna==2.7
jmespath==0.9.4
Pillow==5.2.0
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.1
six==1.13.0
urllib3==1.23
